+++ lib/rubocop/cop/style/string_literals.rb
@@ -22,7 +22,17 @@ module RuboCop
           children = node.children
           return unless children.all? { |c| c.str_type? || c.dstr_type? }
 
-          quote_styles = children.map { |c| c.loc.begin.source }.uniq
+          quote_styles = children.map { |c| c.loc.begin }
+
+          quote_styles = if quote_styles.all?(&:nil?)
+                           # For multi-line strings that only have quote marks
+                           # at the beginning of the first line and the end of
+                           # the last, the begin and end region of each child
+                           # is nil. The quote marks are in the parent node.
+                           [node.loc.begin.source]
+                         else
+                           quote_styles.map(&:source).uniq
+                         end
           if quote_styles.size > 1
             add_offense(node, :expression, MSG_INCONSISTENT)
           else